+++ RELEASE-NOTES.md
@@ -1,6 +1,12 @@
 ddth-dao release notes
 ======================
 
+0.3.0 - 2014-11-08
+------------------
+- New package(s) `com.github.ddth.dao.nosql.*` to support NoSQL-DAOs.
+- Support NoSQL engine: `Apache Cassandra`.
+
+
 0.2.1 - 2014-11-04
 ------------------
 - Moved class `DbcHelper` to package `com.github.ddth.dao.jdbc`.